Tests - Split HRRRGrib tests.
Separate between class level and integration tests with GRIB files.
